Added active link on page
added active link code to layout.jade.  Need to find CSS code to change
the active coloring...right now it shows up as a white box....all the
new sites need is a title to match against in the layout.jade document.
just copy the li code from a previous entry and change the title to
whatever you want your page to be.